minor bug fixed

+++ src/boot/ksdk1.1.0/warp-kl03-ksdk1.1-boot.c
@@ -462,7 +462,7 @@ disableSPIpins(void)
 
 
 void
-enableI2Cpins(uint16_t pullupValue)
+enableI2Cpins(uint8_t pullupValue)
 {
 	CLOCK_SYS_EnableI2cClock(0);
